util: add SIGHUP handling
In the era of remote work the different terminal connections are used for
testing and debugging linuxptp based solutions. Particularly one of the
setup is based on xterm when a set are HW platforms remotely configured and
linuxptp based SW started. When restart of xterm sessions happens the
linuxptp based SW does not receive any of SIGINT/SIGQUIT/SIGTERM signal and
as result left in unpredictable state which causes automation tests failure
after restart and required reboot to recover.

The Xterm will notify its child process with a SIGHUP ("hangup") before it
exits, hence add registration of SIGHUP handler to fix an issue.

Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>