Prevent client ports getting stuck in the UNCALIBRATED state.
When using the 'servo_offset_threshold' configuration option, the
internal servo state will transition from the SERVO_LOCKED to the
SERVO_LOCKED_STABLE state once the synchronization stabilizes.

However, if the network changes in such a way that the selected remote
server appears on a different port, then the client port will never
leave the UNCALIBRATED state.  Fix the issue by dispatching the
appropriate event in the "stable" servo state.

Reported-by: Luigi 'Comio' Mantellini <luigi.mantellini@gmail.com>
Signed-off-by: Richard Cochran <richardcochran@gmail.com>