clock: Accept new UTC offset after leap second.
When the master updated the UTC offset and leap flags in an announce
message after a leap second, only the new flags were accepted for
synchronization of the local clock, which caused a one-second error to
appear until a state decision event updated the UTC offset. Before
commit 14a97dc5da47 ("Remove redundant test on the UTC flags."), the UTC
offset was taken directly from the clock's timePropertiesDS.

Move the update of the UTC offset from clock_update_slave() to
clock_update_time_properties() to enable a synchronous update of the
flags and offset on any announce message.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>
Fixes: 14a97dc5da47 ("Remove redundant test on the UTC flags.")