clock: Clear leap flags after leap second.
When operating as a grandmaster, clear the programmed leap second and
update the UTC offset when the first announce message after the leap
second is sent. This guarantees the update happens in +/- 2 announce
messages around the UTC midnight as required by the specification and
removes the responsibility from the external process which programmed
the leap second with the GRANDMASTER_SETTINGS_NP management message.

The code reads the clock directly instead of using a packet timestamp to
not depend on (timing of) sync messages.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>