phc_ctl: replace calculate_offset with clockadj_compare
The phc_ctl program includes an implementation for comparing the PHC
with CLOCK_REALTIME when PTP_SYS_OFFSET ioctls are not supported.

This implementation produces inverted results when compared with the
implementation of the ioctl.

The PTP_SYS_OFFSET ioctls calculate the difference as "CLOCK_REALTIME -
PHC_TIME" while this function calculates "PHC_TIME - CLOCK_REALTIME".
This could produce confusing results if phc_ctl is used on a kernel
which lacks the PTP_SYS_OFFSET ioctls.

Fix this by replacing the bespoke implementation in test_phc with the
clockadj_compare function. That function mimics the interface of
sysoff_measure and correctly computes the time offset between the two
clocks in the same manner as the ioctls do. In addition, it supports
performing multiple samples to reduce inaccuracy.

With this change, the fallback implementation of measurement will match
the behavior of the ioctl based PTP_SYS_OFFSET flows.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>