Add support for binding sockets to virtual clocks.
With the latest kernels it is possible to create virtual PHCs on top of
a free-running physical PHC. In order for the application to get
timestamps captured by the clock which it is controlling, it needs to
bind its sockets to the clock using a new field in the SO_TIMESTAMPING
option.

Extend the interface structure with the vclock index and modify the
transport code to pass it to sk_timestamping_init() to bind the sockets
to the clock.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>