tlv: Add PORT_HWCLOCK_NP.
Add a new command to get the PHC index associated with the port. This
will be needed for phc2sys -a to use the correct PHC for synchronization
if ptp4l is using a virtual clock.

The TLV also contains a flag indicating a virtual clock.

To follow the PORT_PROPERTIES_NP access policy, PORT_HWCLOCK_NP is
limited to the UDS-RW port.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>