util: import port_state_normalize() logic from phc2sys
PMC agents such as phc2sys tend to suppress uninteresting port state
transitions, to avoid running useless code when nothing really changed
in ptp4l.

With the addition of a PMC agent in ts2phc, it becomes desirable to
place that logic in a common implementation file, to avoid duplication.
Choose util.c for lack of a better place. Also add some comments and
rewrite the implementation so that it is a bit more clear. No change
intended in the produced code.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>