ts2phc: instantiate a pmc agent
This introduces the '-a' option in ts2phc, an option inspired from
phc2sys that puts the clocks in "automatic" mode. In this mode, ts2phc
listens, as a PMC, to port state change events from ptp4l, and detects
which port state machine, if any, has transitioned to PS_SLAVE. That
port's clock will become the synchronization reference for the clock
hierarchy described by ts2phc.

The use case is a multi-switch DSA setup with boundary_clock_jbod, where
there is only one grandmaster, connected to one switch's port. The other
switches, connected together through a PPS signal, must adapt themselves
to this time reference, while the switch connected to the GM must not be
synchronized by ts2phc because it is already synchronized by ptp4l.

Graphically, the setup looks like this:

      +---------------------------------------------------------------+
      | LS1028A                           /dev/ptp0 (unused)          |
      |               +------------------------------+                |
      |               |      DSA master for Felix    |                |
      |               |(internal ENETC port 2: eno2))|                |
      |  +------------+------------------------------+-------------+  |
      |  | Felix embedded L2 switch          /dev/ptp1             |  |
      |  |                             PPS source, sync target     |  |
      |  | +--------------+   +--------------+   +--------------+  |  |
      |  | |DSA master for|   |DSA master for|   |DSA master for|  |  |
      |  | |  SJA1105 1   |   |  SJA1105 2   |   |  SJA1105 3   |  |  |
      |  | |(Felix port 1)|   |(Felix port 2)|   |(Felix port 3)|  |  |
      +--+-+--------------+---+--------------+---+--------------+--+--+

        /dev/ptp2                /dev/ptp3                /dev/ptp4
 PPS sink, sync reference   PPS sink, sync target     PPS sink, sync target
+-----------------------+ +-----------------------+ +-----------------------+
|   SJA1105 switch 1    | |   SJA1105 switch 2    | |   SJA1105 switch 3    |
+-----+-----+-----+-----+ +-----+-----+-----+-----+ +-----+-----+-----+-----+
|sw1p0|sw1p1|sw1p2|sw1p3| |sw2p0|sw2p1|sw2p2|sw2p3| |sw3p0|sw3p1|sw3p2|sw3p3|
+-----+-----+-----+-----+ +-----+-----+-----+-----+ +-----+-----+-----+-----+
   ^
   |
GM connected here

In the described state, the Felix embedded L2 switch is statically
configured to emit a 1PPS signal on a sort of "simplex bus", and SJA1105
switches 1 to 3 are statically configured to record it in their
respective time bases. Because the GM is connected to one of the ports
of SJA1105 switch 1, that port is in PS_SLAVE and is the source of
ts2phc time. So the embedded L2 switch as well as the other SJA1105
switches synchronize to the SJA1105.

When the GM migrates to a port belonging to a different PHC (say sw2p0),
the PPS distribution tree remains unchanged but the port reported by
ptp4l as being in PS_SLAVE changes. So ts2phc must reconfigure itself to
use /dev/ptp3 as time reference, and the other PHC devices turn into
target clocks.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>