ts2phc: allow PHC PPS sources to be synchronized
Now that we are registering a clock even for the PPS source when it
supports that (i.e. when it is a PHC), introduce a new API to retrieve
its clock in order to add timestamps to it.

The timestamps of periodic output transitions are not really known.
We configure the kernel to emit periodic output and then never hear back
from that PHC again. We just assume that it emits periodic output as
promised, and that each pulse is emitted at the beginning of each
second. We rely on the kernel to wake us up when PPS sinks have an extts
event to report, and we know who generated that - the PPS source, of
course. So then we proceed to read the PPS source's PHC time, and round
that to what is the most plausible integer second in its time base.
We believe that to be the 'timestamp'. This is fed into the servo
algorithm.

The PHC PPS source can be synchronized to the extts events of a PHC
slave, when in automatic mode.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>