sk: Handle EINTR when waiting for transmit timestamp.
If ptp4l received a signal in poll() waiting for a transmit timestamp,
multiple (possibly confusing) error messages are logged before exit.

If poll() returns with EINTR, call it once again to get the timestamp
and avoid logging the errors. Don't call it in a loop to avoid getting
stuck in case the timestamp is lost and the signal is repeated before
the poll timeout can be reached.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>