port: don't clear fault if link is down
When ptp4l detects a fault, the application stops for a short time until
the fault_reset_interval has passed.

Since commit 3b82f227ce96 ("clock: Monitor the link status using a RT
netlink socket."), the clock state machine will not try to restart a port
when link is down. This makes sense because a port without link will
obviously just immediately fault again.

However, if a user configures fault_reset_interval with ASAP this behavior
differs. The ASAP value is a special keyword to request clearing the fault
as soon as possible. In this case, the fault is cleared in port.c
immediately as its detected.

Doing so ignores the link status, and thus restarts even when link is down.
This leads to further timeouts and faults. Rather than let these faults
happen just follow the same behavior pattern as the timer-based fault
clearing. Don't clear the fault when link is down, instead wait until link
is restored.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>