Strip Parallel Redundancy Protocol (PRP) trailer
Strip the IEC62439-3 PRP trailer if it is present
to support PTP over PRP.
The implementation is very pedantic about
trailing bytes and will indicate bad message if the
PRP trailer bytes are present when parsing the PTP message.

The PRP trailer is normally removed by the PRP implementation
and invisible to the rest of the stack.
If a PRP HW implementation is used the trailer won't be there.
When the SW PRP implementation introduction in linux kernel 5.9
is used, the trailer will be there.
In short, PRP uses dual lines to send the same information,
LAN A and LAN B, the first arriving packet is handled and
the duplicate from the other interface is dropped.
PTP over PRP can't mix the information from the 2 lines
(path delay might be different) so it is needed to run
a ptp4l instance on each interface instead of on the combined
prp0 interface created by the kernel.
Hence, the PRP trailer will still be there in PTP packages
received directly from the interfaces, the PRP stack hasn't
had the chance to remove it
(which it would if traffic is read from prp0).

Signed-off-by: Magnus Armholt <magnus.armholt@fi.abb.com>