config: allow fractional freq_est_interval
The freq_est_interval value is documented as the time interval over which
to estimate the ratio between the local and peer clock frequencies when
operating in free_running mode.

It is documented as a power of two in seconds, similar to the
summary_interval and the various intervals between packets.

The configuration forces this value to be positive, meaning that fractional
estimation intervals are not supported.A This appears to have been the case
since commit 0513c3b15b8a ("config: Apply more strict input validation to
almost all config file options").

Because of this, when operating in free_running mode it is not possible to
get an output of the estimated ratio faster than once a second, even if
your configuration has a sync rate that is higher.

The commit which enforced this positive boundary did not provide any
justification for why the freq_est_interval cannot be negative. Fix the
specification in the configuration table to correctly allow such intervals.

Fixes: 0513c3b15b8a ("config: Apply more strict input validation to almost all config file options")
Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>