Improve efficiency of nullf servo synchronization
The nullf servo can now enter the SERVO_LOCKED_STABLE state by
transitioning first to the SERVO_LOCKED state when the offset is less than
the set value for offset_threshold. If offset_threshold is not set, the
SERVO_LOCKED state can be entered when the offset is less than or equal to
the set value for step_threshold.

Signed-off-by: Rahul Rameshbabu <rrameshbabu@nvidia.com>