raw: Use BPF filter based on tcpdump syntax.
Replace the hand written BPF code with something that has been created
by tcpdump based on a filter rule. This has the advantage that it can be
extended/ modified based text syntax and is safer to extend in regard to
jump labels.
The generated asm/ BFP code is longer by one opcode because the "and"
operation from VLAN and non-VLAN comparison is not optimized/ merged as
it is the case in the hand-written code.
Also provide two structs/ filters which either filter for the generic or
event PTP packet instead of accessing the struct directly and changing
the jump opcode.

The result is less readable if it comes to offsets. If this is an issue
than it could be optimized with some kind of pre-processor.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>