Add refclock_sock servo.
Add a second servo that provides samples to other processes in order to
control the clock. The chrony SOCK refclock uses a Unix domain socket
instead of a shared memory segment.

The main advantage over the NTP SHM refclock is better security as the
socket can be located in a directory with restricted access, while the
shared memory segment (using a predictable key) can be created by
untrusted users or applications if they can run before ptp4l/phc2sys and
chronyd/ntpd, similarly to the issue with binding to an unprivileged
TCP/UDP port.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>