ts2phc: Fix generic pps source when tai offset is not set in OS
This patch adds support for the leapfile in the generic pps source.
Such implementations prevents situtation when TAI offset is not set in the
system (which is true when there is no NTP daemon running that sets it) from
setting the UTC time in the disciplined PHC.

Leapfile and lstab will only be used when the tai offset read from system is 0.

Signed-off-by: Maciek Machnikowski <maciek@machnikowski.net>
Reviewed-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Maciek Machnikowski <maciek@machnikowski.net>