Set controlField to zero in message headers
IEEE1588-2019 and 802.1AS-2020 both require the field to be 0 in
transmitted messages except when the "version 1 hardware option" is set
(IEEE1588-2019 C.4.2) and only with IPv4.  Since ptp4l isn't padding
the UDP messages to 124 bytes as required for compatibility with the
"V.1 Hardware" implementations, assume that we're not dealing with such
hardware, do not implement the option as defined in C.4.2.

Since the default behaviour changes, in principle this could break
things.  It is similar to 2a2532d66121 ("Bump to IEEE 1588-2019 version")
however, where ptp4l switches the version field to the IEEE1588-2019
value.

[ RPC: Remove every last trace of the control field codes. ]

Signed-off-by: Andrew Zaborowski <andrew.zaborowski@intel.com>
Signed-off-by: Richard Cochran <richardcochran@gmail.com>