ntpshm: Invalidate SHM data before releasing the servo
The SHM is not cleared upon servo release when exiting the tool. This
leaves the last update in the shared memory region marked as valid.

Users may not expect such behavior. If the configuration changed, or the
tool was restarted to reset the state after an unexpected clock step
(e.g. after system suspend), an incorrec timestamp might be accepted by
the consumer

To prevent such behavior - invalidate the SHM data when releasing
the SHM servo.

v2: Fix commit message

Signed-off-by: Maciek Machnikowski <maciek@machnikowski.net>