Flush timer file descriptors after expiration.
After the timer expiration, just after the poll(), we can read fron the
timer file descriptor a 64bit value that represents the number of
expirations that have occured. This value is useful for debug. In
addition, when we use a periodic timer the timer will be locked until
the read of this value.