Split announce/sync_rx timer rearm on expiration.
During a timer handling we should avoid to call timerfd_settime() on a
different timer because this call will invalidate the file descriptor,
resulting in a EGAIN when the other timer code will performe the read().

The ANNOUNCE and RX SYNC timer code just doesn't follow this rule. The
solution is to split the rearm code.