Use periodic timer for TX SYNC timeout.
In order to obtain a precise TX SYNC frequency we need to use a periodic
timer facility instead to reconfigure the timer on every expiration.

The solution consists to configure the it_interval field of the
itimerspec structure (see timerfd_create(2) man page) only when the port
goes into Master role and avoiding to call timerfd_settime() on timer
expiration.