power profile: Fix regression in the default configuration file.
The recently added power profile introduced a regression when reading
the default configuration file:

  $ ./ptp4l -f configs/default.cfg -m -q
  0xFFFFFFFF is an out of range value for option power_profile.2011.grandmasterTimeInaccuracy at line 44
  failed to parse configuration file configs/default.cfg

Root cause is the fact that the value 0xffffffff exceeds the range of
the signed, four byte integer used in the configuration logic.

Considering that the power profile's "Inaccuracy" field uses
0xffffffff as a special value meaning "unknown", expand the range
of these options to include -1 as that special value.

Fixes: 7059a05a3fb2182e851217c8872cf47126ec4c4c ("Introduce the power profile.")
Reported-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Richard Cochran <richardcochran@gmail.com>