Resolve false hybrid_e2e warning
When delay_mechanism is set to "E2E" mode in the [global] section,
it is applied only to non-UDS ports as UDS ports will override the
delay_mechanism with "DM_AUTO". Still the UDS ports will be checked
for hybrid_e2e and generate a false warning of: "hybrid_e2e only
works with E2E".

Update the check so it would only cover non-UDS ports.

Signed-off-by: Eyal Itkin <eitkin@nvidia.com>
Reviewed-by: Rahul Rameshbabu <rrameshbabu@nvidia.com>
Reviewed-by: Jacob Keller <jacob.e.keller@intel.com>