ts2phc: Fix memory leak on initial error path.
The ts2phc program keeps track of the allocated configuration in its
"private" data structure, duly freeing it during the cleanup method.
However, the assignment of the priv.cfg field occurs after many
possible calls to the ts2phc_cleanup() function, resulting in a memory
leak.  Fix the issue by assigning the field prior to the first
invocation of the cleanup method.

Signed-off-by: Richard Cochran <richardcochran@gmail.com>