Avoid switching PHC when phc_index is negative
A Boundary Clock might choose to switch a PHC when jbod is active,
and the transition will be based on the port's phc_index. However,
when using the slave event monitoring + free_running the phc_index
is -1, thus causing a fault in the transition.

Hence, avoid switching the PHC in case phc_index is negative.

Signed-off-by: Eyal Itkin <eitkin@nvidia.com>
Reviewed-by: Rahul Rameshbabu <rrameshbabu@nvidia.com>