Use the 802.1AS peer delay computation when transportSpecific is 1
If the transportSpecific value is configured to be 1, compute t3c using
the 802.1AS math instead of the 1588 math.

The 1588 defined computation using variable names from this project is:
    D = [(t4 - t1) - (t3 - t2) - c1 - c2]/2

The existing code combines the corrections into a variable t3c which we
can get the value of by isolating t3, c1 and c2:
    t3c = t3 + c1 + c2

The 802.1AS defined computation is:
    D = [(t4 - t1) - (t3 - t2) + c1 - c2]/2

Again, isolating t3, c1 and c2 for 802.1AS:
    t3c = t3 + c2 - c1

This has been tested against the MOTU Switch based on the KSZ9567 with
non-zero correction fields based on the 802.1AS equation as well as MOTU
audio interfaces that don't utilize the correction fields.

Signed-off-by: Dylan Robinson <dylan_robinson@motu.com>